requirements.txt
Django==1.10.1
django-cors-headers==1.1.0
django-tastypie==0.13.3
djangorestframework==3.4.6
python-dateutil==2.5.3
python-mimeparse==1.5.2
requests==2.11.1
six==1.10.0
